Add fitness caching
The RMS misfit is always positive, so the calculateRMSMisfit function
now avoids re-running the forward model if the target model already has
a positive fitness. If the fitness is negative, this means it needs to
be recalculated and then the forward model calculations are performed.

A model's fitness is invalidated when a model is mutated or when it is
created as a child.

This should produce a small performance improvement, with a bigger
improvement in inversions with a large number of elitist clones.